Added all functionalities with small bugs
