+++ frontend/src/app/invite/[token]/goal/goalTable/page.jsx
@@ -1,6 +1,10 @@
-import PlayerColl from '../playerCollection/PlayerColl.jsx'
-import Back from '../../../../components/BackButton/BackButton.jsx'
-import styles from "../page.module.css";
+"use client";
+
+import PlayerColl from '../../playerCollection/PlayerColl.jsx';
+import Back from '../../../../../components/BackButton/BackButton.jsx';
+import { useSearchParams, useParams } from "next/navigation";
+import styles from "../../page.module.css";
+
 
 let players = [
   {
@@ -104,7 +108,11 @@ let players = [
     jerseyNum: "2"
   },
 ];
-export default function PlayerOut() {
+export default function GoalTable() {
+  const searchParams = useSearchParams();
+  const { token } = useParams();
+  let scoreUs = searchParams.get("scoreUs");
+  let scoreOpp = searchParams.get("scoreOpp");
 
   return (
       <>
@@ -112,8 +120,8 @@ export default function PlayerOut() {
             <Back/>
         </div>
         <div className={styles.containerColumn}>
-            <h1 className={styles.pageHeader}>Udskiftninger-Ud</h1>
-            <PlayerColl route={"/invite/token/subs/subsTable"} data={players}/>
+            <h1 className={styles.pageHeader}>Mål</h1>
+            <PlayerColl route={`/invite/${token}`} data={players} scores={[parseInt(scoreUs)+1, scoreOpp]}/>
         </div>
       </>
     )
