+++ frontend/src/app/invite/[token]/playerCollection/PlayerColl.jsx
