+++ frontend/src/app/invite/[token]/playerCollection/PlayerRow.jsx
