+++ frontend/src/app/invite/[token]/shot/page.jsx
@@ -1,13 +1,18 @@
+"use client";
+
 import AppCard from '../../../../components/AppCard/AppCard.jsx'
 import TimerCard from '../../../../components/TimerCard/TimerCard.jsx'
 import Back from '../../../../components/BackButton/BackButton.jsx'
 import Score from '../../../../components/Score/Score.jsx'
 import Clock from '../../../../components/Clock/Clock.jsx'
+import { useSearchParams, useParams } from "next/navigation";
 import styles from "../page.module.css";
 
 export default function Shot() {
-  let scoreUs = 14;
-  let scoreOpp = 3;
+  const searchParams = useSearchParams();
+  const { token } = useParams();
+  let scoreUs = searchParams.get("scoreUs");
+  let scoreOpp = searchParams.get("scoreOpp");
 
   return (
         <>
@@ -42,14 +47,16 @@ export default function Shot() {
                     iconColor={"rgb(29, 158, 117)"}
                     iconAlt={"Skud på mål icon"}
                     body={"På mål"}
-                    route={"/invite/token/shot/shotTable"}
+                    scores={[scoreUs, scoreOpp]}
+                    route={`/invite/${token}/shot/shotTable`}
                 />
                 <AppCard 
                     icon={"/xApp.svg"}
                     iconColor={"rgb(232, 67, 12)"}
                     iconAlt={"Skud uden for mål icon"}
                     body={"Uden for mål"}
-                    route={"/invite/token/shot/shotTable"}
+                    scores={[scoreUs, scoreOpp]}
+                    route={`/invite/${token}/shot/shotTable`}
                 />
             </div>
         </>
